1、JFrog的webhook删除事件存入数据库，异步定时任务处理 2、删除制品时，同时删除制品对应的布局坐标顶点 3、仓库清空制品按钮

+++ folib-web-core/src/main/java/com/veadan/folib/components/webhook/DockerWebhooksEventProvider.java
@@ -292,8 +292,17 @@ public class DockerWebhooksEventProvider extends BaseWebhookEventsProvider {
                     return true;
                 }
                 if (Files.exists(repositoryPath)) {
-                    RepositoryFiles.delete(repositoryPath);
-                    result = checkNotExists(repositoryPath);
+                    if (2 == type) {
+                        //定时任务触发，直接删除
+                        RepositoryFiles.delete(repositoryPath);
+                        result = checkNotExists(repositoryPath);
+                    } else {
+                        //webhook触发，保存至数据库，异步定时任务处理
+                        WebhookEventsLog webhookEventsLog = WebhookEventsLog.builder().eventType(webhook.getEventType()).eventRepositoryId(artifactData.getRepoKey()).storageId(storageId).repositoryId(repositoryId).artifactName(name)
+                                .artifactPath(path).sha256Checksum(artifactData.getSha256()).size(artifactData.getSize()).status(WebhookEventsStatusEnum.INIT.getStatus()).build();
+                        webhookEventsLogService.saveWebhookEventsLog(webhookEventsLog, type);
+                        result = true;
+                    }
                 }
             }
         } catch (Exception ex) {
